% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_discount.R
\name{calc_discount}
\alias{calc_discount}
\title{Calculate discount for project}
\usage{
calc_discount(workpackages, initcosting, discount_db, snf = FALSE, dlf = FALSE)
}
\arguments{
\item{workpackages}{work package data (dataframe)}

\item{initcosting}{binary, initial costing for project?}

\item{discount_db}{previous discount applied to project, from REDCap}

\item{snf}{binary, is this an SNF project?}

\item{dlf}{binary, is this a project with DLF support? (unused)}
}
\description{
Calculate discount for project
}
