% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes.R
\name{concat_notes}
\alias{concat_notes}
\title{concatenate external note fields with division/workpackage}
\usage{
concat_notes(notes, header_sep = "\\n\\n", collapse = "\\n\\n")
}
\arguments{
\item{notes}{list of notes from get_notes}

\item{header_sep}{separater between workpackage and note}

\item{collapse}{line break between notes}
}
\description{
concatenate external note fields with division/workpackage
}
