% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costing_info.R
\name{costing_info}
\alias{costing_info}
\title{extract generic costing information from redcap data}
\usage{
costing_info(data, metadata)
}
\arguments{
\item{data}{dataset from redcap}

\item{metadata}{dataset from redcap (i.e. meta$metdata)}
}
\description{
extract generic costing information from redcap data
}
