% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rc_link.R
\name{create_rc_link}
\alias{create_rc_link}
\title{create a link to redcap for a specific costing}
\usage{
create_rc_link(record, costing, token)
}
\arguments{
\item{record}{record_id to link to}

\item{costing}{costing number}

\item{token}{redcap token}
}
\description{
create a link to redcap for a specific costing
}
