% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snf_expense_proportions_table.R
\name{create_snf_expense_proportions_table}
\alias{create_snf_expense_proportions_table}
\title{create SNF table for entering proportions of hours per year}
\usage{
create_snf_expense_proportions_table(expenses, years)
}
\arguments{
\item{years}{number of years}

\item{wp}{workpackage data}
}
\description{
create SNF table for entering proportions of hours per year
}
