% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_ids.R
\name{event_ids}
\alias{event_ids}
\title{REDCap event IDs (for use in creating links to REDCap events within projects)}
\usage{
event_ids()
}
\description{
REDCap event IDs (for use in creating links to REDCap events within projects)
}
