% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expenses_prep.R
\name{expenses_prep}
\alias{expenses_prep}
\title{prepare expenses data}
\usage{
expenses_prep(data, meta)
}
\arguments{
\item{data}{expenses data.frame}

\item{meta}{metadata list}
}
\description{
prepare expenses data
}
