% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pdf.R
\name{gen_pdf}
\alias{gen_pdf}
\title{generate PDF costing report}
\usage{
gen_pdf(output, inputs, copy_html = FALSE)
}
\arguments{
\item{output}{output file name}

\item{inputs}{list of data passed to rmarkdown (see Rmd file for definitions)}

\item{copy_html}{binary, whether to copy the html file to the output directory (primarily for testing purposes)}
}
\description{
generate PDF costing report
}
