% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_pdf.R
\name{gen_pdfq}
\alias{gen_pdfq}
\title{generate PDF costing report via quarto}
\usage{
gen_pdfq(output, inputs, copy_html = FALSE)
}
\arguments{
\item{output}{output file name}

\item{inputs}{list of data passed to rmarkdown (see Rmd file for definitions)}

\item{copy_html}{binary, whether to copy the html file to the output directory (primarily for testing purposes)}
}
\description{
generate PDF costing report via quarto
}
