% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get data from redcap}
\usage{
get_data(record, costing, token)
}
\arguments{
\item{record}{record_id from redcap}

\item{costing}{costing number (event number) from redcap}

\item{token}{API token}
}
\description{
get data from redcap
}
