% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ftes.R
\name{get_ftes}
\alias{get_ftes}
\title{Extract FTE data from downloaded data}
\usage{
get_ftes(d, meta, include = TRUE)
}
\arguments{
\item{d}{downloaded data}

\item{meta}{metadata}

\item{include}{include FTE data in the costing?}
}
\description{
Extract FTE data from downloaded data
}
