% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workpackage_data.R
\name{get_generic_df}
\alias{get_generic_df}
\title{extract generic workpackage data}
\usage{
get_generic_df(d, meta)
}
\arguments{
\item{d}{dataframe}
}
\description{
extract generic workpackage data
}
