% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes.R
\name{get_notes}
\alias{get_notes}
\title{extract external note fields}
\usage{
get_notes(data, meta)
}
\arguments{
\item{data}{list of dataframes from get_data}

\item{meta}{metadata from get_meta}
}
\description{
extract external note fields
}
