% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_testdata_r1}
\alias{get_testdata_r1}
\title{Record TEST, costing 1, from REDCap for unit testing purposes}
\format{
A list of data frames, mostly empty
}
\usage{
get_testdata_r1()
}
\description{
The datasets are generated and save via the code in data-raw/generate_test_datasets.R
}
