% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workpackage_data.R
\name{get_workpackage_data}
\alias{get_workpackage_data}
\title{get all work package data from redcap data}
\usage{
get_workpackage_data(d, meta)
}
\arguments{
\item{d}{data from redcap}

\item{meta}{metadata from redcap}
}
\description{
This functions wraps the functions \code{\link{get_wp_df}} and \code{\link{get_generic_df}} to create a harmonised structure.
}
\seealso{
get_wp_df, get_generic_df
}
