% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_to_dataframe.R
\name{info_to_dataframe}
\alias{info_to_dataframe}
\title{convert the info object into the format output to excel for admins}
\usage{
info_to_dataframe(info)
}
\arguments{
\item{info}{the info object - the output from \code{costing_info()}}
}
\description{
convert the info object into the format output to excel for admins
}
