% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_notes.R
\name{notes_filter}
\alias{notes_filter}
\title{filter notes based on selected workpackages}
\usage{
notes_filter(notes, selected_workpackages)
}
\arguments{
\item{notes}{list of notes from get_notes}

\item{selected_workpackages}{which packages have been selected}
}
\description{
filter notes based on selected workpackages
}
