% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhead.R
\name{overhead}
\alias{overhead}
\title{calculate overhead from workpackages}
\usage{
overhead(discount)
}
\arguments{
\item{discount}{the output from the calc_discount function}
}
\description{
calculate overhead from workpackages
}
