% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_record_inputs.R
\name{record_costing_exists}
\alias{record_costing_exists}
\title{check that a record and costing exists in a project}
\usage{
record_costing_exists(record, costing, token)
}
\arguments{
\item{record}{record_id to check}

\item{costing}{costing number to check}

\item{token}{redcap API token}
}
\description{
check that a record and costing exists in a project
}
