% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_record_inputs.R
\name{record_costings_exist}
\alias{record_costings_exist}
\title{check that costings exist}
\usage{
record_costings_exist(datalist)
}
\arguments{
\item{datalist}{list of dataframes from redcap}
}
\description{
check that costings exist
}
