% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_record_inputs.R
\name{record_meta_enough}
\alias{record_meta_enough}
\title{check that sufficient metadata is available}
\usage{
record_meta_enough(datalist)
}
\arguments{
\item{datalist}{list of dataframes from redcap}
}
\description{
check that sufficient metadata is available
}
