% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_costing_app.R
\name{run_costing_app}
\alias{run_costing_app}
\title{run the shiny app}
\usage{
run_costing_app(opts = list())
}
\arguments{
\item{opts}{list of options to pass to shiny::shinyAppDir}
}
\description{
run the shiny app
}
