% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_expenses_for_admin}
\alias{select_expenses_for_admin}
\title{select relevant expenses variables for admin}
\usage{
select_expenses_for_admin(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
select relevant expenses variables for admin
}
