% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_expenses_for_pdf}
\alias{select_expenses_for_pdf}
\title{select relevant expenses variables for PDF}
\usage{
select_expenses_for_pdf(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
select relevant expenses variables for PDF
}
