% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_for_pdf}
\alias{select_for_pdf}
\title{select relevant workpackage variables for PDF}
\usage{
select_for_pdf(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
select relevant workpackage variables for PDF
}
