% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_fte_for_admin}
\alias{select_fte_for_admin}
\title{select relevant FTE variables for admin}
\usage{
select_fte_for_admin(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
select relevant FTE variables for admin
}
