% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_cost_table.R
\name{snf_cost_table}
\alias{snf_cost_table}
\title{function to convert from total costs per workpackage to cost per year per workpackage}
\usage{
snf_cost_table(workpackages, proportions)
}
\arguments{
\item{workpackages}{typically the output from selected_workpackages}

\item{proportions}{the matrix of proportions of hours per year per workpackage}
}
\description{
function to convert from total costs per workpackage to cost per year per workpackage
}
