% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_expenses_cost_table.R
\name{snf_expenses_cost_table}
\alias{snf_expenses_cost_table}
\title{function to convert from total costs per workpackage to cost per year per workpackage}
\usage{
snf_expenses_cost_table(expenses, proportions)
}
\arguments{
\item{proportions}{the matrix of proportions of hours per year per workpackage}

\item{workpackages}{typically the output from selected_workpackages}
}
\description{
function to convert from total costs per workpackage to cost per year per workpackage
}
