% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_iict_yearly_budget.R
\name{snf_iict_yearly_budget}
\alias{snf_iict_yearly_budget}
\title{Fill out the SNF IICT budget file, distributing hours over years}
\usage{
snf_iict_yearly_budget(hours, expenses, info, debug = FALSE)
}
\arguments{
\item{hours}{A data frame with the hours per year (result from snf_cost_table)}

\item{expenses}{A data frame with the expenses per year (result from snf_expenses_cost_table)}

\item{info}{A list with the project information}

\item{debug}{return extra logging information for debugging}
}
\description{
SNF has a specific budget form for IICT budgets.
}
