% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_option.R
\name{specific_option}
\alias{specific_option}
\title{Given a specific variable, return the label of the option selected}
\usage{
specific_option(metadata, data, var)
}
\arguments{
\item{metadata}{metadata/data dictionary from redcap}

\item{data}{dataset from redcap}

\item{var}{the variable of interest}
}
\description{
Given a specific variable, return the label of the option selected
}
