% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_.R
\name{summarize_by_wp}
\alias{summarize_by_wp}
\alias{summarize_by_div}
\title{Summarize by particular level}
\usage{
summarize_by_wp(data)

summarize_by_div(data)
}
\arguments{
\item{data}{A data frame with columns Service, wp, wp_lab, desc, Hours, Units, Rate, Cost}
}
\description{
summarize_by_wp summarizes by work package
}
\details{
summarize_by_div summarizes by division
}
