% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Calculate totals}
\usage{
totals(
  workpackages,
  expenses,
  discount,
  discount_chf = 0,
  overhead,
  fte,
  snf,
  rate,
  dlf = FALSE
)
}
\arguments{
\item{workpackages}{A tibble with the workpackages}

\item{expenses}{A tibble with the expenses}

\item{discount}{A tibble with the discount}

\item{discount_chf}{Discount value (numeric)}

\item{overhead}{A tibble with the overhead}

\item{fte}{fte cost data}

\item{snf}{A boolean indicating whether the project is an SNF project}

\item{rate}{A string indicating the rate (internal, external for-profit, external non-profit)}

\item{dlf}{A boolean indicating whether the project is a DLF project or not (unused)}
}
\description{
pass the output from the relevant functions to this function to calculate the totals.
}
