% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_codes.R
\name{wp_codes}
\alias{wp_codes}
\title{Get projectfacts work package codes}
\usage{
wp_codes(metadata)
}
\arguments{
\item{metadata}{A data frame with columns val, lab. I.e. meta$metadata}
}
\description{
Get projectfacts work package codes
}
