% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overhead.R
\name{overhead}
\alias{overhead}
\title{calculate overhead from workpackages}
\usage{
overhead(workpackages)
}
\arguments{
\item{workpackages}{A tibble with the workpackages}
}
\description{
calculate overhead from workpackages
}
