% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snf_iict_yearly_budget.R
\name{snf_iict_yearly_budget}
\alias{snf_iict_yearly_budget}
\title{Fill out the SNF IICT budget file, distributing hours over years}
\usage{
snf_iict_yearly_budget(hours, info, output)
}
\arguments{
\item{hours}{A data frame with the hours per year (result from snf_cost_table)}

\item{info}{A list with the project information}

\item{rates}{A data frame with the rates}
}
\description{
SNF has a specific budget form for IICT budgets.
}
