% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Calculate totals}
\usage{
totals(workpackages, expenses, discount, overhead, internal, dlf = FALSE)
}
\arguments{
\item{workpackages}{A tibble with the workpackages}

\item{expenses}{A tibble with the expenses}

\item{discount}{A tibble with the discount}

\item{overhead}{A tibble with the overhead}

\item{internal}{A boolean indicating whether the project is internal or not}

\item{dlf}{A boolean indicating whether the project is a DLF project or not (unused)}
}
\description{
pass the output from the relevant functions to this function to calculate the totals.
}
